Update : 로그인 구현 전 오류 수정

+++ back-end/youandi/src/main/java/com/ssafy/youandi/service/UserService.java
@@ -1,10 +1,10 @@
 package com.ssafy.youandi.service;
 
-import com.ssafy.youandi.entity.user.SignupRequest;
+import com.ssafy.youandi.dto.SignupRequest;
 
 
 public interface UserService {
     //회원가입
-    public int singUp(SignupRequest signupRequest) throws Exception;
+    public Long singUp(SignupRequest signupRequest) throws Exception;
 
 }
